reorg of weekplan
