+++ src/server/api/routers/dinner.ts
@@ -2,7 +2,6 @@ import { z } from "zod";
 
 import { createTRPCRouter, publicProcedure } from "~/server/api/trpc";
 import { type DinnerWithTags } from "../../../utils/types";
-import { type Dinner } from "@prisma/client";
 import { getFirstAvailableDay } from "../../../utils/dinner";
 
 export const dinnerRouter = createTRPCRouter({
