+++ src/views/WeekPlan/DinnerSlot.tsx
@@ -1,60 +1,21 @@
+import { useDraggable } from "@dnd-kit/core";
 import { type Dinner } from "@prisma/client";
-import { api } from "../utils/api";
-import { cn } from "../lib/utils";
-import { getWeekPlan } from "../utils/dinner";
-import { useDraggable, useDroppable } from "@dnd-kit/core";
+import { cn } from "../../lib/utils";
+import { api } from "../../utils/api";
 
-export const WeekPlan = () => {
-  const dinnersQuery = api.dinner.dinners.useQuery();
+type DinnerSlotProps = { dinner?: Dinner };
 
-  const weekPlan = getWeekPlan(dinnersQuery.data?.dinners);
+export const DinnerSlot = ({ dinner }: DinnerSlotProps) => {
+  if (!dinner) {
+    return <div className="h-12 rounded-md border"></div>;
+  }
 
-  const days = [
-    "Monday",
-    "Tuesday",
-    "Wednesday",
-    "Thursday",
-    "Friday",
-    "Saturday",
-    "Sunday",
-  ];
-
-  return (
-    <div className="flex flex-col items-end justify-start bg-gray-50 p-6 dark:bg-gray-900">
-      <h2 className="mb-8 text-right text-xl font-bold">Week Plan</h2>
-      <div className="w-full space-y-4 text-right">
-        {days.map((day, index) => (
-          <Day key={day} day={day} dinner={weekPlan[index]} />
-        ))}
-      </div>
-    </div>
-  );
+  return <DinnerSelected dinner={dinner} />;
 };
 
-type DayProps = { day: string; dinner?: Dinner };
+type DinnerSelectedProps = { dinner: Dinner };
 
-const Day = ({ day, dinner }: DayProps) => {
-  const { isOver, setNodeRef } = useDroppable({
-    id: day,
-  });
-
-  return (
-    <div
-      className={cn(
-        "flex flex-col rounded border px-2 py-2",
-        isOver && "bg-green-400",
-      )}
-      ref={setNodeRef}
-    >
-      <h3 className="mb-2 mr-1 text-xs">{day}</h3>
-      <Dinner dinner={dinner} />
-    </div>
-  );
-};
-
-type DinnerProps = { dinner?: Dinner };
-
-const Dinner = ({ dinner }: DinnerProps) => {
+const DinnerSelected = ({ dinner }: DinnerSelectedProps) => {
   const utils = api.useUtils();
   const unselectDinnerMutation = api.dinner.unselect.useMutation({
     onMutate: (input) => {
@@ -89,17 +50,13 @@ const Dinner = ({ dinner }: DinnerProps) => {
   });
 
   const { attributes, listeners, setNodeRef, transform } = useDraggable({
-    id: dinner?.id ?? "",
+    id: dinner?.id,
   });
   const style = transform
     ? {
         transform: `translate3d(${transform.x}px, ${transform.y}px, 0)`,
       }
     : undefined;
-
-  if (!dinner) {
-    return <div className="h-12 rounded-md border"></div>;
-  }
 
   return (
     <div
